refactor(pgn): move ParsePGN and types to chess/pgn sub-package
PGNTags, result constants (ResultWhiteWins/BlackWins/Draw/Ongoing), and
ParsePGN (now pgn.Parse) move to the new chess/pgn sub-package.  The
chess package retains PGN() and generation helpers (determineResult,
buildMoveText, writeTag, wrapLines, tagValue) and imports chess/pgn for
the shared types.

This mirrors the standard Go pattern: pure data types and parsing logic
live in a sub-package that has no dependency on the engine's unexported
state, while the engine-coupled generator stays in chess/ and imports
that sub-package.
