+++ chess/pgn_test.go
@@ -5,6 +5,7 @@ import (
 	"testing"
 
 	"github.com/RchrdHndrcks/gochess/chess"
+	chesspgn "github.com/RchrdHndrcks/gochess/chess/pgn"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -14,7 +15,7 @@ func TestPGN(t *testing.T) {
 		c, err := chess.New()
 		require.NoError(t, err)
 
-		pgn := c.PGN(chess.PGNTags{})
+		pgn := c.PGN(chesspgn.PGNTags{})
 
 		assert.Contains(t, pgn, `[Event "?"]`)
 		assert.Contains(t, pgn, `[Site "?"]`)
@@ -23,9 +24,9 @@ func TestPGN(t *testing.T) {
 		assert.Contains(t, pgn, `[White "?"]`)
 		assert.Contains(t, pgn, `[Black "?"]`)
 		assert.Contains(t, pgn, `[Result "*"]`)
-		parsedTags, parsedMoves, parseErr := chess.ParsePGN(pgn)
+		parsedTags, parsedMoves, parseErr := chesspgn.Parse(pgn)
 		require.NoError(t, parseErr)
-		assert.Equal(t, chess.ResultOngoing, parsedTags.Result)
+		assert.Equal(t, chesspgn.ResultOngoing, parsedTags.Result)
 		assert.Empty(t, parsedMoves)
 	})
 
@@ -33,7 +34,7 @@ func TestPGN(t *testing.T) {
 		c, err := chess.New()
 		require.NoError(t, err)
 
-		tags := chess.PGNTags{
+		tags := chesspgn.PGNTags{
 			Event: "Test Tournament",
 			Site:  "Internet",
 			Date:  "2026.03.17",
@@ -63,11 +64,11 @@ func TestPGN(t *testing.T) {
 
 		require.True(t, c.IsCheckmate())
 
-		pgn := c.PGN(chess.PGNTags{})
+		pgn := c.PGN(chesspgn.PGNTags{})
 
-		parsedTags, parsedMoves, parseErr := chess.ParsePGN(pgn)
+		parsedTags, parsedMoves, parseErr := chesspgn.Parse(pgn)
 		require.NoError(t, parseErr)
-		assert.Equal(t, chess.ResultWhiteWins, parsedTags.Result)
+		assert.Equal(t, chesspgn.ResultWhiteWins, parsedTags.Result)
 		assert.Equal(t, []string{"e2e4", "e7e5", "f1c4", "b8c6", "d1h5", "g8f6", "h5f7"}, parsedMoves)
 	})
 
@@ -75,7 +76,7 @@ func TestPGN(t *testing.T) {
 		c, err := chess.New()
 		require.NoError(t, err)
 
-		pgn := c.PGN(chess.PGNTags{})
+		pgn := c.PGN(chesspgn.PGNTags{})
 
 		for _, line := range strings.Split(pgn, "\n") {
 			assert.LessOrEqual(t, len(line), 80, "line exceeds 80 characters: %s", line)
@@ -86,7 +87,7 @@ func TestPGN(t *testing.T) {
 		c, err := chess.New()
 		require.NoError(t, err)
 
-		tags := chess.PGNTags{
+		tags := chesspgn.PGNTags{
 			Event: "He said \"hello\"",
 			Site:  "path\\to\\file",
 		}
@@ -96,87 +97,10 @@ func TestPGN(t *testing.T) {
 		assert.Contains(t, pgn, "[Site \"path\\\\to\\\\file\"]")
 
 		// Roundtrip: parse back and verify unescaped values.
-		parsedTags, _, err := chess.ParsePGN(pgn)
+		parsedTags, _, err := chesspgn.Parse(pgn)
 		require.NoError(t, err)
 		assert.Equal(t, "He said \"hello\"", parsedTags.Event)
 		assert.Equal(t, "path\\to\\file", parsedTags.Site)
-	})
-}
-
-func TestParsePGN(t *testing.T) {
-	t.Run("Simple PGN", func(t *testing.T) {
-		pgn := `[Event "Test"]
-[Site "Internet"]
-[Date "2026.03.17"]
-[Round "1"]
-[White "Alice"]
-[Black "Bob"]
-[Result "1-0"]
-
-1. e2e4 e7e5 2. f1c4 b8c6 3. d1h5 g8f6 4. h5f7 1-0
-`
-		tags, moves, err := chess.ParsePGN(pgn)
-		require.NoError(t, err)
-
-		assert.Equal(t, "Test", tags.Event)
-		assert.Equal(t, "Internet", tags.Site)
-		assert.Equal(t, "2026.03.17", tags.Date)
-		assert.Equal(t, "1", tags.Round)
-		assert.Equal(t, "Alice", tags.White)
-		assert.Equal(t, "Bob", tags.Black)
-		assert.Equal(t, "1-0", tags.Result)
-
-		expectedMoves := []string{"e2e4", "e7e5", "f1c4", "b8c6", "d1h5", "g8f6", "h5f7"}
-		assert.Equal(t, expectedMoves, moves)
-	})
-
-	t.Run("PGN with comments and variations", func(t *testing.T) {
-		pgn := `[Event "?"]
-[Result "*"]
-
-1. e2e4 {Best move} e7e5 (1... d7d5 2. e4d5) 2. g1f3 $1 b8c6 *
-`
-		tags, moves, err := chess.ParsePGN(pgn)
-		require.NoError(t, err)
-
-		assert.Equal(t, "?", tags.Event)
-		assert.Equal(t, "*", tags.Result)
-
-		expectedMoves := []string{"e2e4", "e7e5", "g1f3", "b8c6"}
-		assert.Equal(t, expectedMoves, moves)
-	})
-
-	t.Run("PGN with semicolon comments", func(t *testing.T) {
-		pgn := `[Event "?"]
-[Result "*"]
-
-1. e2e4 e7e5 ; this is a comment
-2. g1f3 b8c6 *
-`
-		_, moves, err := chess.ParsePGN(pgn)
-		require.NoError(t, err)
-
-		expectedMoves := []string{"e2e4", "e7e5", "g1f3", "b8c6"}
-		assert.Equal(t, expectedMoves, moves)
-	})
-
-	t.Run("Empty PGN", func(t *testing.T) {
-		pgn := `[Event "?"]
-[Site "?"]
-[Date "?"]
-[Round "?"]
-[White "?"]
-[Black "?"]
-[Result "*"]
-
-*
-`
-		tags, moves, err := chess.ParsePGN(pgn)
-		require.NoError(t, err)
-
-		assert.Equal(t, "?", tags.Event)
-		assert.Equal(t, "*", tags.Result)
-		assert.Empty(t, moves)
 	})
 }
 
@@ -190,14 +114,14 @@ func TestPGNRoundtrip(t *testing.T) {
 			require.NoError(t, c.MakeMove(m))
 		}
 
-		tags := chess.PGNTags{
+		tags := chesspgn.PGNTags{
 			Event: "Roundtrip Test",
 			White: "W",
 			Black: "B",
 		}
 		pgn := c.PGN(tags)
 
-		parsedTags, parsedMoves, err := chess.ParsePGN(pgn)
+		parsedTags, parsedMoves, err := chesspgn.Parse(pgn)
 		require.NoError(t, err)
 
 		assert.Equal(t, "Roundtrip Test", parsedTags.Event)
